Fixed Euler integration
Euler integration had the wrong derivative.  We were missing an -1j from
the superoperator representing the derivative of the density matrix.

In this commit we set the total time steps to be equal to 100.  You can
see how a large time step is a poor choice for Euler integration.
Increasing the number of time steps (decreasing dt) returns us to the
expected accuracy of linear integration methods.