code doc and kraus operators
Form kraus operators for the reduced system dynamics.  Seemed to work
fine but we had an issue with normalization.  Completeness was a little
off...where this stems from...not sure...

Tried to start code clean up so we can eventually port to larger
systems.

Next on the agenda is map to super methods.  Represent density matrix in
column stacked form.  Represent super operater as a matrix on 4^{N}
space.